requirements.txt
django-watermark>=0.1.8